docs: update all documentation for new CLI, logger, and reporter
This commit brings all documentation in line with the recent CLI overhaul,
removing references to deprecated options and adding the new ones.

Global changes:
- Replace --verbose, --output-format, --quiet with
  --report, --json, --log-level, --log-file
- Update all command examples to use the new options
- Add sections explaining the three independent output streams
  (findings, logs, JSON reports)
- Mention fork/thread tracing, eBPF backend, and test coverage

Files updated:
- README.md: Quick start, examples, status, logging section
- ROADMAP.md: Mark completed items (fork/thread, logger/reporter, tests)
- THREAT_MODEL.md: Reflect new capabilities and limitations
- FAQ.md: Add CLI options explanation, update Q&A
- ARCHITECTURE.md: Add logger, reporter, backends, output separation
- docs/security_checks.md: Update dynamic analysis examples, CLI flags
- docs/extending_detectors.md: Mention reporter integration
- docs/integration_examples.md: Update JSON and logging examples
- docs/use_cases.md: Add reporter integration notes
- docs/index.html: Update landing page with new features and CLI

+++ examples/custom_analyzers/credit_card_analyzer.cpp
@@ -52,7 +52,7 @@ runtimexray::FindingList CreditCardAnalyzer::analyze(const runtimexray::Evidence
                 ++end;
             }
             if (digits >= 13 && digits <= 19) {
-                std::string card = chunk.substr(pos, digits);
+                std::string card = chunk.substr(pos, static_cast<std::size_t>(digits));
                 findings.emplace_back(
                     runtimexray::FindingSeverity::High,
                     "Credit card number found in memory",
